App.js cleanup
